bug: wrong order of fiat values for token transactions, calc of weth values wrong

+++ src/main/scala/info/graphsense/Model.scala
@@ -39,7 +39,7 @@ case class TokenConfiguration(
     token_address: Array[Byte],
     standard: String,
     decimals: Int,
-    decimal_divisor: Int,
+    decimal_divisor: Long,
     peg_currency: Option[String]
 )
