+++ src/main/scala/info/graphsense/TokenTransfers.scala
@@ -19,7 +19,7 @@ class TokenTransfers(spark: SparkSession) {
       hexstr_to_bytes("0xdAC17F958D2ee523a2206206994597C13D831ec7"),
       "erc20",
       6,
-      pow(10, 6).intValue(),
+      pow(10, 6).longValue(),
       Some("USD")
     ),
     TokenConfiguration(
@@ -27,7 +27,7 @@ class TokenTransfers(spark: SparkSession) {
       hexstr_to_bytes("0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48"),
       "erc20",
       6,
-      pow(10, 6).intValue(),
+      pow(10, 6).longValue(),
       Some("USD")
     ),
     TokenConfiguration(
@@ -35,7 +35,7 @@ class TokenTransfers(spark: SparkSession) {
       hexstr_to_bytes("0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2"),
       "erc20",
       18,
-      pow(10, 18).intValue(),
+      pow(10, 18).longValue(),
       Some("ETH")
     )
   )
